backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pydantic
requests
authlib
pyjwt
sqlmodel
